+++ apps/mcp-server/RELEASES.md
@@ -15,3 +15,11 @@ Based on:
 - Speakeasy CLI 1.768.2 (2.889.1) https://github.com/speakeasy-api/speakeasy
 ### Generated
 - [mcp-typescript v0.0.2] apps/mcp-server
+
+## 2026-06-02 01:21:18
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.769.2 (2.892.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [mcp-typescript v0.0.3] apps/mcp-server